platform-core/data-generators/sports-generator/requirements.txt
requests
Faker
flask
kafka-python